codex: consolidate to ACP-only, stream native frontend events
Per direction to stop maintaining two transports: drop the CLI (codex exec)
backend and make codex() ACP-only. Merge the ACP client into codex.py and
remove codex_acp.py — one path, one file.

Frontend contract (investigated against oo-chat + connectonion-ts): the SDK's
chat-item-mapper renders a fixed event vocabulary. The old custom
io.log('codex_event', ...) matched nothing and would not render. Fix: convert
Codex's ACP session/update into the SDK's NATIVE events —
  - tool_call  (stable tool_id, name)      -> running tool card
  - tool_result(same tool_id, done/error)  -> its completion
and permission requests already map to agent.io.request_approval (rendered as
an approval card, answer flows back). So NO frontend or SDK change is needed.

- codex(): ACP-only; sandbox/model become codex-acp -c overrides; approval
  governs per-action session/request_permission (manual->io, auto->allow,
  no io->deny). Returns the same envelope.
- last_message accumulates only agent_message_chunk (thoughts excluded).
- reader thread + permission handler fail safe (deny) so a raise can't hang
  the turn; utf-8 errors='replace'.

Tests: test_codex_tool.py rewritten for the ACP-only tool incl. native-event
assertions; real e2e (test_real_codex.py) rewritten to drive the real
codex-acp binary. Validated end-to-end against codex-acp 0.16.0 (handshake +
clean auth-error). test_codex_acp.py folded in; codex_acp.py removed.
